Improved Project Structure * Updated linting configuration for backend * Removed redundant styling * Updated component names for better comprehension
