+++ frontend/history.html
@@ -9,7 +9,6 @@
   </head>
   <body>
     <div id="app"></div>
-
-    <script type="module" src="./src/history.js"></script>
+    <script type="module" src="./src/components/history/history.js"></script>
   </body>
 </html>
