+++ frontend/index.html
@@ -17,7 +17,7 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/cart.js"></script>
+    <script type="module" src="/src/app.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
   </body>
 </html>
