+++ frontend/src/components/cart/Cart.vue
@@ -1,0 +1,38 @@
+<script setup>
+import CartModal from "./CartModal.vue";
+import { $viewCart } from "./cart";
+import { useStore } from "@nanostores/vue";
+const viewCart = useStore($viewCart);
+</script>
+
+<template>
+  <div class="Cart">
+    <!-- Cart Icon -->
+    <div class="cart-icon" @click="$viewCart.set(true)">
+      <svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"
+        x="0px" y="0px" width="40px" height="36px" viewBox="0 0 40 36" style="enable-background: new 0 0 40 36"
+        xml:space="preserve">
+        <g id="Page-1_4_" sketch:type="MSPage">
+          <g id="Desktop_4_" transform="translate(-84.000000, -410.000000)" sketch:type="MSArtboardGroup">
+            <path id="Cart" sketch:type="MSShapeGroup" class="st0"
+              d="M94.5,434.6h24.8l4.7-15.7H92.2l-1.3-8.9H84v4.8h3.1l3.7,27.8h0.1
+                c0,1.9,1.8,3.4,3.9,3.4c2.2,0,3.9-1.5,3.9-3.4h12.8c0,1.9,1.8,3.4,3.9,3.4c2.2,0,3.9-1.5,3.9-3.4h1.7v-3.9l-25.8-0.1L94.5,434.6" />
+          </g>
+        </g>
+      </svg>
+    </div>
+    <!-- Cart Modal -->
+    <cart-modal v-if="viewCart" />
+  </div>
+</template>
+
+<style scoped>
+svg {
+  fill: none;
+  stroke: black;
+  /* Change the color of the icon */
+  width: 50px;
+  /* Change the size of the icon */
+  height: 50px;
+}
+</style>
