+++ frontend/support.html
@@ -10,6 +10,6 @@
   <body>
     <div id="app"></div>
 
-    <script type="module" src="./src/support.js"></script>
+    <script type="module" src="./src/components/contact-support/support.js"></script>
   </body>
 </html>
